`ucm init` instead of `ucm` in quickstart
When I was following the quickstart, when I ran `ucm` in `Step 2: Create your Unison codebase`, then I got the output:
```
$ ucm
  ⚠️
  No codebase exists in /my-path/homedir Run `ucm init` to create one, then try again!
```
So, I ran `ucm init` as it suggested:
```
$ ucm init
  ⚠️
  Initializing a new codebase in: /my-path/homedir
```
I could be confused, but figured I'd throw up this PR if that's the correct thing to do, thanks! Let me know if you have any questions, I'm new to Unison.