add minimal command reference
This PR is intended as a quick pass at creating a placeholder for the command reference, until #60 can be completed. It documents a "minimal" set of UCM commands—just the commands that are needed to support self-discovery in UCM itself (e.g., `help`, `find`, `docs`), as well as `pull` (due to the need to pull a base library before doing anything substantive with Unison), [`quit`](https://stackoverflow.blog/2017/05/23/stack-overflow-helping-one-million-developers-exit-vim/), and some placeholder sections to avoid broken internal links.