Point to unison-nix repo in docs
The [nixpkgs](https://github.com/nixos/nixpkgs) derivation for Unison is
outdated and unfortunately PRs to update it aren't getting merged. This
means that people who install Unison via Nix are getting an outdated
version (currently M1l) that is known to be buggy and can have issues
pulling in the base library.

After some discussion with @aryairani I created
[unison-nix](https://github.com/ceedubs/unison-nix/) as an attempt to
make it as easy as possible for people to install an up-to-date version
of Unison via Nix. It's not a perfect solution, and as it is in a
separate repo it is bound to lag behind updates in the main Unison repo.
But I think that it's better than the current solution of an outdated
nixpkgs derivation.